requirements.txt
browser-use
python-dotenv
playwright
streamlit
